+++ packages/asset-swapper/src/utils/market_operation_utils/types.ts
@@ -70,6 +70,7 @@ export enum ERC20BridgeSource {
     ShibaSwap = 'ShibaSwap',
     AaveV2 = 'Aave_V2',
     Compound = 'Compound',
+    Synapse = 'Synapse',
     // BSC only
     PancakeSwap = 'PancakeSwap',
     PancakeSwapV2 = 'PancakeSwap_V2',
@@ -115,7 +116,7 @@ export type SourcesWithPoolsCache =
 export enum CurveFunctionSelectors {
     None = '0x00000000',
     exchange = '0x3df02124',
-    exchange_underlying = '0xa6417ed6',
+    exchange_underlying = '0xa6417ed6', // exchange_underlying(int128 i, int128 j, uint256 dx, uint256 min_dy)
     get_dy_underlying = '0x07211ef7',
     get_dx_underlying = '0x0e71d1b9',
     get_dy = '0x5e0d443f', // get_dy(int128,int128,uint256)
@@ -130,7 +131,7 @@ export enum CurveFunctionSelectors {
     // Smoothy
     swap_uint256 = '0x5673b02d', // swap(uint256,uint256,uint256,uint256)
     get_swap_amount = '0x45cf2ef6', // getSwapAmount(uint256,uint256,uint256)
-    // Nerve BSC, Saddle Mainnet
+    // Nerve BSC, Saddle Mainnet, Synapse
     swap = '0x91695586', // swap(uint8,uint8,uint256,uint256,uint256)
     calculateSwap = '0xa95b089f', // calculateSwap(uint8,uint8,uint256)
 }